refactor(editor): studio flow hooks 분리
